+++ nzshm_model/logic_tree/branch.py
@@ -6,13 +6,13 @@ from abc import ABC, abstractmethod
 from dataclasses import dataclass, field
 from functools import reduce
 from operator import mul
-from typing import TYPE_CHECKING, Sequence, Union
+from typing import TYPE_CHECKING, Sequence
 
 if TYPE_CHECKING:
     from .logic_tree_base import BranchSet, FilteredBranch, LogicTree
 
 
-@dataclass(frozen=True)
+@dataclass
 class Branch(ABC):
     """
     Abstract baseclass for logic tree branches